Split discount types (#12927)
