+++ saleor/graphql/discount/types/__init__.py
@@ -1,0 +1,13 @@
+from .discounts import OrderDiscount
+from .sales import Sale, SaleChannelListing, SaleCountableConnection
+from .vouchers import Voucher, VoucherChannelListing, VoucherCountableConnection
+
+__all__ = [
+    "OrderDiscount",
+    "Sale",
+    "SaleChannelListing",
+    "SaleCountableConnection",
+    "Voucher",
+    "VoucherChannelListing",
+    "VoucherCountableConnection",
+]
